requirements.txt
boto3==1.35.34
botocore==1.35.34
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
iniconfig==2.0.0
jmespath==1.0.1
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.2
six==1.16.0
tscribe==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webvtt-py==0.5.1
